backend/requirements.txt
requests
google-cloud-bigquery
pandas
pyarrow
db-dtypes
pandas-gbq